refactor : error handling 구조 리팩토링
