+++ src/main/java/gift/GlobalExceptionHandler.java
@@ -1,6 +1,12 @@
-package gift.product.error;
+package gift;
 
+import gift.member.util.AuthenticationFailedException;
+import gift.product.error.AlreadyExistsException;
+import gift.product.error.NotFoundException;
+import java.util.HashMap;
+import java.util.Map;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -28,5 +34,13 @@ public class GlobalExceptionHandler {
     public String handleValidationException(MethodArgumentNotValidException ex, Model model){
         model.addAttribute("errorMessage", "유효성 검사 실패: " + ex.getBindingResult().getFieldError().getDefaultMessage());
         return "error";
+    }
+
+    @ExceptionHandler(AuthenticationFailedException.class)
+    @ResponseStatus(HttpStatus.CONFLICT)
+    public ResponseEntity<Map<String, String>> handleAuthenticationFailedException(AuthenticationFailedException e) {
+        Map<String, String> errorResponse = new HashMap<>();
+        errorResponse.put("error", e.getMessage());
+        return new ResponseEntity<>(errorResponse, HttpStatus.UNAUTHORIZED);
     }
 }
